backend/pyproject.toml
[project]
name = "holyread-backend"
version = "1.0.0"
description = "HolyRead V1 — Bíblia + Hebraico/Grego original"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic>=2.9",
  "slowapi>=0.1.9",
  "python-dotenv>=1.0",
  "orjson>=3.10",
]

[project.optional-dependencies]
etl = [
  "lxml>=5.3",
  "requests>=2.32",
]
translate = [
  "deep-translator>=1.11",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
